notebook/requirements.txt
numpy
scipy
jupyter
sklearn
pandas
prompt-toolkit==1.0.15
newspaper3k
